N°5749 - PR review plus plus - enhance curl options refactor
