+++ core/jsoncollector.class.inc.php
@@ -142,17 +142,7 @@ abstract class JsonCollector extends Collector
 				$aDataGet = [];
 			}
 			$iSynchroTimeout = (int)Utils::GetConfigurationValue('itop_synchro_timeout', 600); // timeout in seconds, for a synchro to run
-
-			$aRawCurlOptions = Utils::GetConfigurationValue('curl_options', array(CURLOPT_SSLVERSION => CURL_SSLVERSION_SSLv3));
-			$aCurlOptions = array();
-			foreach ($aRawCurlOptions as $key => $value) {
-				// Convert strings like 'CURLOPT_SSLVERSION' to the value of the corresponding define i.e CURLOPT_SSLVERSION = 32 !
-				$iKey = (!is_numeric($key)) ? constant((string)$key) : (int)$key;
-				$iValue = (!is_numeric($value)) ? constant((string)$value) : (int)$value;
-				$aCurlOptions[$iKey] = $iValue;
-			}
-			$aCurlOptions[CURLOPT_CONNECTTIMEOUT] = $iSynchroTimeout;
-			$aCurlOptions[CURLOPT_TIMEOUT] = $iSynchroTimeout;
+			$aCurlOptions = Utils::GetCurlOptions($iSynchroTimeout);
 
 			//logs
 			Utils::Log(LOG_DEBUG, 'Source aDataGet: '.json_encode($aDataGet));
