+++ core/restclient.class.inc.php
@@ -1,7 +1,7 @@
 <?php
 // Copyright (C) 2014 Combodo SARL
 //
-//   This application is free software; you can redistribute it and/or modify	
+//   This application is free software; you can redistribute it and/or modify
 //   it under the terms of the GNU Affero General Public License as published by
 //   the Free Software Foundation, either version 3 of the License, or
 //   (at your option) any later version.
@@ -116,15 +116,7 @@ class RestClient
 		$aData['json_data'] = json_encode($aOperation);
 		$sUrl = Utils::GetConfigurationValue('itop_url', '').'/webservices/rest.php?login_mode=form&version='.$sVersion;
 		$aHeaders = array();
-		$aRawCurlOptions = Utils::GetConfigurationValue('curl_options', array());
-		$aCurlOptions = array();
-		foreach ($aRawCurlOptions as $key => $value) {
-			// Convert strings like 'CURLOPT_SSLVERSION' to the value of the corresponding define i.e CURLOPT_SSLVERSION = 32 !
-			// Make sure that the actual constant exists to prevent warnings/errors in case of typo or when php-curl is not loaded
-			$iKey = ((!is_numeric($key)) && defined((string)$key)) ? constant((string)$key) : (int)$key;
-			$iValue = ((!is_numeric($value)) && defined((string)$value)) ? constant((string)$value) : (int)$value;
-			$aCurlOptions[$iKey] = $iValue;
-		}
+		$aCurlOptions = Utils::GetCurlOptions();
 		$response = Utils::DoPostRequest($sUrl, $aData, '', $aHeaders, $aCurlOptions);
 		$aResults = json_decode($response, true);
 		if (!$aResults) {
